fix: Issues when PG username is 'cloudquery' (#371)
Before this fix, cloudquery misbehaved when the PG username was named 'cloudquery'.
It created tables in the 'cloudquery' schema instead of the 'public' schema.
This happened because the search path was  '"$user", public', which means it depended on the username.

This fix, upon connection to PG, will change serach_path to 'public' only, i.e. not dependent on username (which is way more predictable).

This won't impact us explicitly accessing the 'cloudquery' schema via 'cloudquery.<table-name>', which is what we do when we actually need the 'cloudquery' schema.