feat: Using json everywhere apart from yaml for the user
We are taking a similar approach to k8s to use json internally
for marshalling/unmarshalling and yaml only for using facing stuff.

yaml parsers are much more complex and also have tons of vulnerabilities
so we want to use json everywhere where there is a machine reading those configurations.