feat: Make NewSourceClient (and dest) as one interface (#208)
#### Summary

I think I finally nailed the Client interface. This will remove
completely the need for "PluginManager" in the CLI and yet another
abstraction layer and prop drilling with WithOptions duplicates for
every abstraction layer.

Also, the SDK is aware of `registry`, `path` and `version` so I really
think it should be here. it turns the CLI intro real CLI finally.

Co-authored-by: Herman Schaaf <hermanschaaf@gmail.com>