chore(ci): Trigger renovate after SDK release (#266)
This is a small optimization to how we do SDK updates in the monorepo. At the moment renovate runs every 15 minutes to handle dependencies updates. We run it every 15 minutes to ensure speedy updates to the SDK.
With this PR we'll run it immediately after the SDK has been released.
This should also allow us to save some build minuted by reducing the renovate schedule to 1 hour for example

---