fix: Add custom log reader implementation to fix hang on long log lines (#263)
This fixes an issue where the CLI would hang indefinitely when it encountered a log line from a plugin that was longer than the buffer size allowed for by `bufio.Scanner`. The fix uses a wrapper around `bufio.NewReader` to skip lines that are too long to be parsed, printing an error in the log with the first 1000 characters to help with identification.