feat: Support application level protocol message. (#294)
This will let us make breaking changes to the protocol while ensuring users get the right message on what todo. e.g update plugin or cli (client).

This is needed in general and also is a pre-requisite for the performance issue in destination plugins that will require breaking change in the destination protocol.