fix: Remove descriptions from table docs (#300)
Descriptions are too detailed for the summary table, and are removed in this PR. 

I'm also changing the format of the table a little bit: instead of having child tables in a separate column, I'm experimenting with having them in their own rows but indicated as a child table with an arrow. Not sure this will be the final version, but I think it looks a bit better than what we had before.