fix: Move cqtypes (#357)
This is a split of this PR
https://github.com/cloudquery/plugin-sdk/pull/348 so it will be easier
to review.

Also, Im going to push the move to v2 maybe another week as we dont
really have anything that requires protocol changes for now. even though
it is the right way to do it and will make the cli much easier including
the need to things like that -
https://github.com/cloudquery/cloudquery/pull/3657 as it will be all
handles on protobuf and go versioning side.

This current PR just moves cqtypes to schema and removes enumerating all
types in different locations to make sure when we add a new type tests
will catch if a type is not added to the right place and or not
connected to `TypeValue`.

Also this removes `TimeInterval` which 1) wasn't used 2) it is incorrect
because it used `time.Duration` as `TimeInterval` but it's not a
TimeInterval it's just a duration which is uint64 so there no need for a
special type for that. In PostgreSQL TimeInterval is actually completely
different type that contains timezones and it is not go `time.Duration`