fix(faker): Use `MarshalText` for faker timestamps (#373)
Fixes https://github.com/cloudquery/cloudquery/issues/3627.

**The TLDR** is that we should `MarshalText` when faking time data, as that follows an RFC.

**The DON'T READ part is**:
If we don't do it and default to `String()` that sometimes creates a string that cannot be parsed by `const defaultStringFormat = "2006-01-02 15:04:05.999999999 -0700 MST"`.

You can see an example in https://github.com/cloudquery/cloudquery/issues/3627 that the `2110-10-19 08:53:50.21670207 +0000 UTC m=+2775422040.494306874` string was created.
The reason that the string can't be parsed is because the nano seconds part has 8 digits and not 9, that is `21670207` instead of `216702070`.
The nano seconds string needs to be padded only when adding the `m=+2775422040.494306874` part. This was added in go 1.19 (see https://pkg.go.dev/time#Time.String).
To summarize:
1. `time.Parse("2006-01-02 15:04:05.999999999 -0700 MST", "2110-10-19 08:53:50.21670207 +0000 UTC")` works
2. `time.Parse("2006-01-02 15:04:05.999999999 -0700 MST", "2110-10-19 08:53:50.216702070 +0000 UTC  m=+2775422040.494306874")` works
3. `time.Parse("2006-01-02 15:04:05.999999999 -0700 MST", "2110-10-19 08:53:50.21670207 +0000 UTC  m=+2775422040.494306874")` fails

So technically this could be a bug in go, that needs to pad the nano seconds when adding the `m=+` part, though there is no guarantee on the format of `time.String()`.

I added another test `TestTimestamptzSetString` that shows this behavior. The only way I could find to create a string that has the `m=+` part is using `time.Now()`. It's in the PR to demonstrate the issue and I'll remove it before merging

---