fix: Add -race when running tests (#388)
This adds the `-race` flag to `go test`, so we can detect data races in the tests. There were two instances that I had to fix for this to succeed (not really a bug: just the test accessing a variable that another goroutine is writing to) - for now fixed using a lock around the variable.

Some other test were failing when I ran `go test -race -count 3 ./...` because of global variables being modified in the test. This also fixes this issue.  

The only real functional change is that we now only add a `_cq_id` column if it is not already present.