fix: Pre-aggregate metrics before sending (#396)
This is a short-term fix for https://github.com/cloudquery/cloudquery/issues/3962 while we work on a better solution. It pre-aggregates metric results before sending it from the server, so that only a small map gets sent over the wire. We can do this because we know the CloudQuery CLI right now only makes use of the aggregated data in any case.