feat: Add support for glob matching in config (#398)
This adds wildcard matching to the tables / skip_tables config entries. It also adds the ability to skip child tables, which may sometimes be useful if the child tables are expensive to fetch.

This will be useful to reduce the number of entries needed in config. For example, to fetch all ec2-related resources, but skip vpn gateways:

```
  tables: 
    - "aws_ec2_*"
  skip_tables:
    - "aws_ec2_vpn_gateways"
```

Instead of  20+ entries that it would have required before.

# Behavior
This is how table configs will work after this change:
 - Glob matches will match against both top-level and descendent tables
 - If you want to sync a table, it needs to be explicitly mentioned. This includes descendent (child) tables. This is the one and only breaking change, because previously child tables needed to be excluded from config.
 - Similarly, if you want to skip a table or a descendent table, it needs to be explicitly skipped in the `skip_tables` section, either through a glob match or by using its exact name.

This will require some config updates, but I believe this is a more consistent approach: if you want a table to by synced, include it. If you want it to be skipped, skip it. There are no implicit inclusions anymore.

# Breaking Change

This will not be considered a breaking plugin-sdk change (no change to the protocol), but it *will* be considered a breaking change in all source plugins, where we should bump the major version. Old configs will continue to work, but the behavior will change slightly for some of them, since top-level tables that were explicitly included will now stop fetching their descendent tables by default. 

Relates to https://github.com/cloudquery/cloudquery/issues/3963