fix: Increase GRPC message size limit to 50MiB (#419)
The default was 4MiB, and users have reported running into this limit on occasion. If 50MiB isn't enough, we'll probably need to implement some logic to drop messages at that point since they'll be too big to be useful for many databases anyway. Let's see if increasing the limit to 50MiB solves the problem first though.

https://github.com/cloudquery/cloudquery/issues/4834