feat: Handle resolving of map[string]string to JSON
In fields like 'Tags', 'Annotations' - it is more consistent if empty/null
dictionaries show up as an empty JSON object `{}`, rather than being a postgres "null"
or a JSON "null". Same for slices.

Note that this PR will not handle the case of nested fields - but it should already be an improvement on current behavior.