fix: Use correct error codes (#514)
This is uses correct error codes for gRPC in most places. We use Internal and this is the intedded error code for server errors. This is also used by the grpc server as well but it doesn't mean it is only for the gRPC code. See documentation:

https://pkg.go.dev/google.golang.org/grpc/codes#Code
https://cloud.google.com/apis/design/errors

The right codes for internal error are `Unknown` (The default one) and `Internal`