fix(testing): Some fixes to the ordering for plugin tests (#565)
There was an issue with append-only tests that only wrote 1 resource: the sort order meant that the incorrect resource was sometimes being compared in this case. 

This PR changes the sort order to be by sync time first, then UUID. Some other tests had to be adapted to work with this new ordering as well.