feat: Add local backend for storing cursor state (#569)
A state store allows us to store cursors for https://github.com/cloudquery/plugin-sdk/issues/413. This first iteration introduces the interface and implements it for a local JSON storage backend. After some back and forth I decided to call it `backend`, similar to Terraform. Alternatives considered were `state` (confusing because it's not a single state) and `storage` (confusing because it's similar to `destination`). 

Since the interface is a generic key-value store, we lose the ability to know whether a table is incremental or not, and which columns are used as cursor keys if it is. I'm proposing that we add some fields to the `Table` definition to be able to at least document these aspects, but it's not enforced.

See https://github.com/cloudquery/cloudquery/pull/6336 for an example implementation using this new functionality.