fix: Send resource validation errors to Sentry (#601)
Resource validation errors, like a row missing a PK value, are non-recoverable and need to be fixed by the plugin author. As such, we should know about them when they happen, even though they do not cause a panic in the code. I tried to minimize the number of duplicate messages we send in this case by using a `sync.Map`