fix: Ignore env variables in comments (#625)
- Fixes https://github.com/cloudquery/plugin-sdk/issues/613

This changes the yaml config parsing code to ignore comments entirely. The solution used here is relatively short and clean, but it required some unfortunate hacking because the curly brackets we use for env variable placeholders are not valid yaml, and so any attempt to parse the yaml with them still in place fails (unless we were to write our own parser). 

Simple regex replacement of comments would also be prone to issues, because it wouldn't be aware of the context (maybe the `#` is inside a string, for example). This solution replaces the variables with temporary placeholders, converts each yaml section to json and back to yaml to strip comments, and then replaces the temporary placeholders again before continuing as before.