feat: Support downloading plugins from other cloudquery repos (#632)
Right now, if a plugin is hosted in anywhere on the `cloudquery` org, it can only be downloaded if it is part of the `cloudquery/cloudquery` monorepo. This makes it impossible to temporarily host `cq-source-*` type plugins as other `cloudquery` repos, as can be done with community plugins on other orgs.

This changes the download logic to first check the monorepo, but fall back to the community standard of `cq-source-*` if it's A) `cloudquery` org and B) the monorepo download attempt resulted in a 404. If both fail with 404, the attempt is aborted. Other status codes are retried.

It would have been better if we could know, somehow, when a plugin is hosted in the monorepo or outside it, but right now this is not possible unless we change the spec in some way. 

This change allows us to use the CLI to download https://github.com/cloudquery/cq-source-simple-analytics