feat: Add skip_dependent_tables option (#662)
This adds a flag to the source config that allows users to opt out of syncing dependent tables by default. This has been raised as an issue by users a few times already, most recently [here](https://discord.com/channels/872925471417962546/873606591335759872/1071081863105495178). 

The issue with the current default behavior is that new dependencies can be added with minor versions, which can 1) drastically slow down a sync and 2) create unexpected migrations or new tables. It can also be surprising for new users, when selecting e.g. only the `aws_s3_buckets` table, that they get ~6 tables synced.

This should ideally be the default behavior, but that would be a breaking change. For now, I think we should keep backwards-compatibility but add this flag (and recommend it as best-practice).

Note about naming: `dependent` is the standard American spelling. `dependant` is used in the UK. I went with the American spelling here.