fix: Fix race in dest testing try 3 (#683)
Quite hard to debug locally https://github.com/cloudquery/cloudquery/pull/7819 so sorry for influx of PRs. 

My current theory is that the db is blocked on number of connections.

The next thing that I might be trying is to not to do those tests concurrently as it can cause weird things in databases and just do those sequentially.