feat(multiplex): Detect duplicated clients (#723)
Related to https://github.com/cloudquery/plugin-sdk/issues/713. This PR detects duplicated client IDs, sends a sentry report when that happens and also logs a warning.
It not this does not skip those clients (yet). I'll do a follow up based on the sentry data as we can have duplicate clients due to:
1. User errors creating a configuration that generates duplicated clients like in https://github.com/cloudquery/cloudquery/pull/8099
2. Plugins bug, not implementing the ID method correctly

---