fix(types-timestamp): Ensure timestamp is UTC (#716)
Fixes https://github.com/cloudquery/plugin-sdk/issues/700

Tested this on quite a few plugins. Most send `time.Time` in UTC already (`nil` location).
The only impact I've seen is on this AWS resource:
https://github.com/cloudquery/cloudquery/blob/380bcddff9406d12b7de343dd4dd002c95b46f0f/plugins/source/aws/resources/services/iam/credential_reports_fetch.go#L66
which for some reason results in a `time.Time` with a location with an empty name, so that would be converted to UTC.

Another impact is on the Stripe plugin that uses `int64`. Since `time.Unix` returns a `time.Time` with local timezone, now it will be converted to `UTC`.

---