fix: Disallow null character in strings per utf8 spec (#797)
This should solve bunch of issues that we have with NullValues and StripNulls logic propagated to all our destination plugins. Apparently [Valid](https://github.com/apache/arrow/pull/35161#discussion_r1170516104) utf8 string can't have null values in the middle of the string so we should just strip those in our type system (it can have at the end or the start but I suggest actually stripping those down as well as it's really unnecessary as far as I know and it exist prob for some archaic reason like sending a string over the wire without any additional encoding/protocol, which we have via arrow).

Once we move sources to arrow we will add the validation/strip on the source side.