fix: Use Go memory allocator for arrow (#810)
Following this short discussion - https://github.com/apache/arrow/issues/35232

It seems we don't really need to use Retain/Release outside the arrow library. We can always bring this back in the future if we would like to experiment if this brings better memory performance then the default go allocator.