feat: Allow testing of more Arrow types (#863)
This adds support for testing (almost) the complete range of Arrow types in destinations.

By default, only the types previously supported by CQTypes are enabled (with the important exception of more uint and int types than before, plus float32). The TestSourceOptions struct allows destinations to declare that they want to test with more column types, including maps, structs, times, etc.

I have not included all the nullable column variations. These are a bit more difficult to do now that we are not using `arrow.Field` directly anymore, plus we can rather do that as a follow-up if and when it becomes necessary to test this.

This PR is a rebased version for v3 that replaces #818.