fix: Fix timer logic in batch writers (#1056)
I believe that the batch timeout should be a guarantee of a message being delayed by no longer than X amount of time. This means that we shouldn't reset the ticker every time a message is received, because then a message could be delayed for a very long time if a slow trickle of messages is coming in. This change fixes this issue and guarantees a message to be flushed within at most X duration.

For more context, see this PR and message https://github.com/cloudquery/plugin-sdk/pull/1055#discussion_r1251726213