fix: Add random suffix to test table names (#1086)
This adds a random suffix to test table names so that it's possible to run the same test more than once in a second without getting name collisions. I chose a random number here because it's simpler, but I could imagine other solutions like an incrementing counter also working with a bit of additional complexity.