feat: Add Unflatten tables method (#1138)
This adds an `UnflattenTables` method to the `Tables` type, the opposite operation of `FlattenTables`.

### Why?
Until now, destination plugins were the only part of the code that needed to decode Arrow schemas to tables, and destinations don't need relational information. Now that docs generation is moving to the CLI, however, we need to be able to reconstruct the original relational structure in order to accurately reflect it in the docs.

Rather than modifying `FlattenTables`, which would not be backwards-compatible, I have opted to add a new method that the CLI can use. With this change in place, I am able to fully generate the docs for the AWS plugin from the CLI, with zero changes.