feat: Add message to package command (#1216)
This adds a required "message" paramater to the package command, which then gets added to package.json. This is for summarizing the changes in a version.

At the moment this flag is required, but we could potentially loosen that restriction if we had a way to use git to track changes between commits.