feat: Support publishing plugins with team and kind metadata set (#1313)
This allows us to package the plugin with team and kind metadata set,
needed for publishing to the new registry and checking usage and limits
against the new cloud API.

The package command no longer needs to take a `kind` argument.

Also adds an `info` command to print this information, useful for
debugging, otherwise it won't be visible anywhere. We can consider
adding this to the `--version` output instead, but that would
technically be breaking?