feat: Adding a batch updater to allow usage updates to be batched (#1326)
This will allows a client to update the usage asynchronously. The updater will only call the API when a configurable number of rows have been updated or a timeout is reached. In addition the updater will flush any remaining rows before closing.

If an error is encountered then an exponential backoff up to a max number of retries or a maximum wait time (relative to the start time of the last query) will be followed.

In addition, if the server replies with a status of `429` and includes a `Retry-After` header, then the client will wait that number of seconds before retrying.


It might be worth considering just using a retry library also e.g. https://github.com/avast/retry-go.