fix: Return clearer error when team name empty or not set (#1354)
Right now the error users get when the team is not set is `failed to
configure quota monitor: failed to get usage: 404 Not Found`. The recent
CLI change in https://github.com/cloudquery/cloudquery/pull/15119 will
ensure that the team should mostly get set, but there can still be cases
where users are in multiple teams and forget to run `cloudquery switch`.
In such a case, a friendlier error would be helpful.