feat: Allow homogeneous data types to be configured
For OpenSearch (and presumably ElasticSearch) arrays (even in JSON objects) are limited to the same data types. This adds a configuration option to the writer test suite to only use the same data type in the JSON example data - allowing the test suite to be used for the OpenSearch plugin. Currently JSON objects in the ElasticSearch and OpenSearch plugins are mapped to strings.