feat: Implement batch sender. (#1995)
## Problem

The DFS scheduler implements concurrent resolution of resources, but only if resources are sent in batches to the top-level `res` channel. Thus, the same plugin syncs a lot quicker simply by batching resources before sending them through the channel.

## Solution this PR implements

```
// BatchSender is a helper struct that batches items and sends them in batches of batchSize or after batchTimeout.
//
// - If item is already a slice, it will be sent directly
// - Otherwise, it will be added to the current batch
// - If the current batch has reached the batch size, it will be sent immediately
// - Otherwise, a timer will be started to send the current batch after the batch timeout
type BatchSender struct {
```

Currently with a batch timeout of `100ms` and batch size of `100`, it manages a 5x improvement on a sync that reproduces the issue brought forward by the community (https://github.com/jeromewir/cq-source-concurrency-childtable-example):

<img width="982" alt="Screenshot 2024-12-03 at 17 17 23" src="https://github.com/user-attachments/assets/7a1f7dc5-4aca-4263-8aa6-0ec10d73cca3">

The PR tries to touch the least amount of scheduler code, and stays away from tricky language constructs as much as possible.