feat: Enable storing _cq_client_id. (#2046)
We need to be able to store `_cq_client_id` on every table for certain
use cases where we need to know where a row came from in terms of
multiplexed clients.

This PR adds the ability for the SDK and all its schedulers to store
`_cq_client_id` on all tables.

I added this feature to the example plugin in the SDK and it does store
it properly:

<img width="861" alt="Screenshot 2025-01-08 at 10 45 04"
src="https://github.com/user-attachments/assets/58705cf2-3e61-4585-bfbc-63abbccef555"
/>

If you want to try this yourself, please note that, very unfortunately,
the simple plugin also stores a `client_id` column that is unrelated
(and doesn't work, since it thinks it's supposed to be an integer).