fix: Require Row count to be greater than 0 (#2259)
#### Summary

Currently when we send usage we iterate over all possible tables. In cases where a table had a non 0 value earlier in the sync it is still included in every subsequent report. This leads us to have large numbers of `row_count = 0` records in our DB

This change will eliminate the SDK from sending these records