feat: Make getColumnChangeSummary public for use in plugins (#2260)
In the ClickHouse destination, we want to add some additional information about table-level changes (sort keys, partition keys) to that provided by `GetChangesSummary`. However right now, that would mean duplicating quite a lot of code, so I would like to suggest that we instead make `GetColumnChangeSummary` public for use in this case and others like it.